+++ lib/src/router/yui_tab_route_state.dart
@@ -1,0 +1,14 @@
+//
+//
+//
+
+/// YuiTabRouterState
+class YuiTabRouterState {
+  final String path;
+  final bool drawerIsOpen;
+
+  YuiTabRouterState({
+    required this.path,
+    required this.drawerIsOpen,
+  });
+}