+++ lib/src/router/yui_tab_router.dart
@@ -5,6 +5,7 @@
 import 'package:flutter/material.dart';
 import 'package:yuix/src/router/yui_nav_router.dart';
 import 'package:yuix/src/router/yui_router_protocol.dart';
+import 'package:yuix/src/router/yui_tab_route_state.dart';
 import 'package:yuix/src/router/yui_tab_state.dart';
 
 /// TabPageBuilder
@@ -19,24 +20,32 @@ class YuiTabRouter implements YuiRouterProtocol {
     String? initialPath,
     required this.pages,
     this.items,
+    this.drawer,
   })  : state = (() {
           // init state
           final paths = pages.keys.toList();
           final path = initialPath ?? paths.first;
-          return ValueNotifier(path);
+          final state = YuiTabRouterState(
+            path: path,
+            drawerIsOpen: false,
+          );
+          return ValueNotifier(state);
         })(),
         initialPath = initialPath ?? pages.keys.toList().first;
 
   /// selected path
-  final ValueNotifier<String> state;
+  final ValueNotifier<YuiTabRouterState> state;
 
   /// members
   final String initialPath;
   final Map<Path, TabPageBuilder> pages;
   final Map<Path, TabBarItemBuilder>? items;
+  final Widget Function()? drawer;
+
+  final drawerKey = GlobalKey<ScaffoldState>();
 
   /// Tab Page ID
-  String get selectedPath => state.value;
+  String get selectedPath => state.value.path;
 
   /// selected index
   int get selectedIndex {
@@ -48,10 +57,13 @@ class YuiTabRouter implements YuiRouterProtocol {
 
   /// Switch Tab
   void select(String path) {
-    if (state.value == path) {
+    if (state.value.path == path) {
       return;
     }
-    state.value = path;
+    state.value = YuiTabRouterState(
+      path: path,
+      drawerIsOpen: state.value.drawerIsOpen,
+    );
   }
 
   /// Switch Tab by index
@@ -60,8 +72,35 @@ class YuiTabRouter implements YuiRouterProtocol {
     select(path);
   }
 
+  /// drawer in
+  void slideIn() {
+    // state
+    final newState = YuiTabRouterState(
+      path: state.value.path,
+      drawerIsOpen: true,
+    );
+    state.value = newState;
+    // controll byself
+    drawerKey.currentState?.openDrawer();
+  }
+
+  /// drawer out
+  void slideOut() {
+    // state
+    final newState = YuiTabRouterState(
+      path: state.value.path,
+      drawerIsOpen: false,
+    );
+    state.value = newState;
+    // controll byself
+    drawerKey.currentState?.closeDrawer();
+  }
+
   /// reset all state
   void resetAllState() {
-    state.value = initialPath;
+    state.value = YuiTabRouterState(
+      path: initialPath,
+      drawerIsOpen: false,
+    );
   }
 }